Drop create_repository variable
This variable was added while still using Terraform 0.12 and allowed us
to include this module in other modules and making creation of a new
repo optional, as using count on a module was not supported.

Since Terraform 0.13 was released we can use count on a module and it's
better to skip the module completely if the create_repository variable
is set to false in the calling module instead of just the
github_repository resource.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>