Expect the default branch to be "main"
This still allows you to create additional branches and set another branch as the default branch, but it turns out that nowadays a repo is always initially created with a main branch.

So if you want to use another default branch, you will first need to create it and then point update the config to point to the new branch.

This PR enables that again as it turned out this doesn't currently work anymore.