feat: Support unmanaged files
Sometimes you want to template a repository where you need interpolation
so creating from a template repository does not work. The content of
these skeleton files shouldn't be managed by Terraform, they're intended
to be written once and then updated outside of Terraform.

This should be a non-breaking change as we add a field to the
`repository_files` variable that defaults to the original behaviour and
include a `moved` block to migrate any existing files from
`github_repository_file.default` to `github_repository_file.managed`.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>