fix!: Cannot create team and repo in same run (#81)
This addresses two issues:
* Fixes a bug where a team cannot be created and passed to this module
  in the same Terraform run because the ID is not known at plan time.
* Changing a team's access level results in resources being deleted and
  created

We now use the team name as a key in our `access` variable. This is safe
as team names must be unique in an organization and are known when a
plan runs. Inside the module we lookup the team ID using a data resource,
resulting in a better user experience and fixing this limitation where you
cannot create teams and pass them to this module in a single Terraform
run because the list passed to the module contained unknown values.

Before we needed to pass in GitHub team IDs to one of the access levels
variables. This is not a future proof solution and means any time we
change an access level, the resource is recreated when it can be updated
via the API. This also allows for easier diffing as you can see the
change in access level instead of resources being deleted and created.

---------

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>