feat: Add `var.merge_strategy` to select merge strategy
In the last major version we defaulted to only enable squash merge; this
was because we saw people choosing different merge types when all
options were available, making for a messy commit history.

This commit adds `var.merge_strategy` which allows the consumer to
select a merge strategy for the repo. This variable defaults to `null`
so that if it's not explicitly set, all merge strategies are enabled by
default, as was the case before.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>