fix(environment-module): Handle null inputs better (#91)
Set `nullable = false` for variables with empty types (e.g. `{}` or
`[]`). We set it for `deployment_policy` so that if someone does pass in
null, we still do the good thing of adding the policy for protected
branches. If they really want to opt out, they can do so with
`protected_branches = false`.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>