fix: Configure `overwrite_on_create` using variable
`var.repository_files` defaults to overwriting files if they already
exist. This functionality should be opt-in, so we default it to `false`
and provide a way for someone to set it to `true`.

With the default settings, the Terraform apply will fail because the
file already exists, at which point it can either be imported, or
`overwrite_on_create` can be set to `true` to proceed.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>