fix!: Remove computable values from `var.deployment_policy` (#94)
* fix!: Remove computable values from `var.deployment_policy`

When creating an environment the default behaviour is to have
`var.deployment_policy`'s `protected_branches` set to true and
`custom_branch_policies` set to false, as these are mutually exclusive.
If you configure `branch_patterns` or `tag_patterns`, you have to set
custom_branch_policies to `true` and `protected_branches` to false -- this
is more cognitive load than we want for our users.

This commit removes the first two values: we keep the default of only
allowing deployments from a protected branch, but, if `branch_patterns`
or `tag_patterns are set`, the module will handle setting the booleans to
make them come into effect.

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>

* docs(readme): update module usage

---------

Signed-off-by: Stephen Hoekstra <shoekstra@schubergphilis.com>
Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>